requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.0.14
djangorestframework==3.16.1
gunicorn==22.0.0
idna==3.11
packaging==26.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
flake8>=7.0,<8.0
